+++ app/build.gradle.kts
@@ -27,10 +27,11 @@ import me.jerryokafor.ihenkiri.Config
 plugins {
     id("me.jerryokafor.ihenkiri.android.application")
     id("me.jerryokafor.ihenkiri.android.application.compose")
+    id("me.jerryokafor.ihenkiri.application.jacoco")
     id("me.jerryokafor.ihenkiri.android.hilt")
     id("me.jerryokafor.ihenkiri.android.navigation")
-    id("me.jerryokafor.ihenkiri.application.jacoco")
 //    id("com.google.android.libraries.mapsplatform.secrets-gradle-plugin")
+    id("jacoco")
 }
 
 android {
@@ -48,12 +49,24 @@ android {
     }
 
     buildTypes {
+        debug {
+            enableUnitTestCoverage = false
+            enableAndroidTestCoverage = false
+        }
         release {
             isMinifyEnabled = false
+            enableUnitTestCoverage = false
+            enableAndroidTestCoverage = false
             proguardFiles(
                 getDefaultProguardFile("proguard-android-optimize.txt"),
                 "proguard-rules.pro",
             )
+        }
+    }
+
+    testOptions {
+        unitTests {
+            isIncludeAndroidResources = true
         }
     }
 }
@@ -103,13 +116,5 @@ dependencies {
     implementation(libs.com.google.accompanist.systemuicontroller)
 
     testImplementation(project(":core:test"))
-    testImplementation(libs.junit4)
-    testImplementation(libs.io.mockk.android)
-    testImplementation(libs.io.mockk.agent)
-
     androidTestImplementation(project(":core:test"))
-    androidTestImplementation(libs.io.mockk.android)
-    androidTestImplementation(libs.io.mockk.agent)
-    androidTestImplementation(libs.androidx.test.ext.junit)
-    androidTestImplementation(libs.androidx.test.espresso.core)
 }
